next, add methods to PhotoFile
