+++ sample.py
@@ -2,7 +2,10 @@
 
 # Tools for making subsets / samples of Red Panda Lineage data
 
-from shared import MEDIA_PATH, PANDA_PATH, ZOO_PATH, SpeciesError, PhotoFile, SectionNameError
+import os
+import sys
+
+from shared import MEDIA_PATH, PANDA_PATH, ZOO_PATH, SpeciesError, PhotoFile, SectionNameError, get_max_entity_count
 
 def sample_photos(animals, photos, size, species, taglist):
     """
@@ -19,7 +22,24 @@ def sample_photos(animals, photos, size, species, taglist):
     ./sample/<species>/metadata/<rpf-id>_photo.<photo.index>.txt
         - photo.url, author, commitdate, link, tags, uncertainty
     """
-    pass
+    max = int(get_max_entity_count())
+    for file_path in [PANDA_PATH]:
+        section = None
+        for section_name in ["pandas"]:
+            if section_name in file_path.split("/"):
+                section = section_name.split("s")[0]   # HACK
+        # Enter the pandas subdirectories
+        for root, dirs, files in os.walk(file_path):
+            for filename in files:
+                path = root + os.sep + filename
+                # print(path)
+                photo_list = PhotoFile(section, path)
+                photo_count = photo_list.photo_count()
+                # TOWRITE: check_species
+                photo_index = 1
+                # TOWRITE: has_tag in PhotoFile
+                # TOWRITE: get_sample method, considers size and counts
+                pass
 
 if __name__ == '__main__':
     # Default settings
